Code: - Create simple VOIDLIST class to store dynamic data. - Adapt IDLIST, OBJECT_LIST to VOIDLIST. - Fixed: Incorrect masks in processMove(). - Fixed: Sometimes user unable to set negative value in EditNumber fields. - Fixed: IniFile does not save empty files. - A lot of various minor changes.
General:
- Added: Import of BUILD maps version 5.
- Fixed: Random crash on PFB (prefab) file load/save.
- Changing first wall of sector should not affect auto-align wall.
- Better highlighted sectors map to map transfer.
- Various minor changes and fixes.

2D Mode:
- Added: Masked wall indication for edit screen.
- Added: Auto-arc building tool.
- Sector shape menu has been renamed to Sector tools menu.
- Door wizard has been moved to Sector tools menu.
- Door wizard ignores wall points on target red wall line
  when you want to create rotate doors.
- Fixed: Data1 sound can play twice for double rotate doors that
  using Link command technique (if there is no Data2 sound defined).
  If you have such doors, you need to disable "send at ON" and
  "send at OFF" checkboxes on both sides.
- Fixed: Flipping sector sometimes may lead to walls corruption.
- Fixed: Splitting sector sometimes may lead to walls duplication.

3D Mode:
- Added: Wall texture auto-alignment to the left. Press [<] to use (still may have issues).
- Fixed: Gradient Shading tool was not working properly.
- Fixed: Frame for hovered sprites has a bit incorrect size for face sprites.
- Fixed: Highlighted sprite extents should not go through angled walls when you drag such sprites.

Preview Mode:
- Added: "Laser" type preview (modern features).
- Added: "Path Sector SPRITE" type preview (modern features).

HUD:
- Added: "Cstat" field in edit dialogs of common objects (sectors, sprites, walls).
- Added: "Detail" field in edit dialog of common sprite.
- Added: Helper function for "Cstat" and "Hi-tag" fields which allows to pick required
  flags instead of typing it manually. Press [F10] to call it. Note: editing flags
  may lead to unexpected results. Please, make sure you understand what each flag
  supposed to do before selecting it.
- Added: Helper function for "Pan angle" and "Wind angle" fields in sector fx dialog
  which automatically sets the angle relative to first wall of the sector.
- Fixed: Some incorrect limits for values in object edit dialogs.

SEQEDIT / QAVEDIT:
- Added: Automatic test map generation when launched in standalone mode, which
  allows to test animations in 3D mode always.

File Browser:
- Added: Confirmation before file deletion.

+++ src_blood/xmpexplo.cpp
@@ -1560,9 +1560,13 @@ char FileBrowser::ItemRemoveDlg(char* pNeedle)
 			chdir(cwd);
 		}
 	}
-	else
+	else if (Confirm("Remove file \"%s\"?", pNeedle))
 	{
 		r = (chmod(newpath, S_IREAD|S_IWRITE) == 0 && unlink(newpath) == 0);
+	}
+	else
+	{
+		return 0;
 	}
 	
 	if (r == 0)
@@ -2354,9 +2358,10 @@ static char getThumbnail_PFB(char* filepath, int nTile, int twh, int thg, int bg
 		////////////////////////////////////
 
 		char tColor = bg;
-		int s, e, wh, hg, zhg, bvisib;
+		int s, e, wh, hg, zhg, bvisib; int32_t* p;
 		int cama = 0, camx = 0, camy = 0, camz;
 		int nSect, nBlank, nAng;
+		IDLIST* pHglt = NULL;
 		sectortype* pSect;
 
 		// insert a large enough sector
@@ -2366,7 +2371,15 @@ static char getThumbnail_PFB(char* filepath, int nTile, int twh, int thg, int bg
 		pSect = &sector[nSect];
 		pSect->ceilingz	= -0x20000;
 		pSect->floorz	=  0x20000;
-
+		
+		if (highlightcnt > 0)
+		{
+			// meh...
+			pHglt = new IDLIST(true);
+			for (s = 0; s < highlightcnt; s++)
+				pHglt->Add(highlight[s]);
+		}
+		
 		getSectorWalls(nSect, &s ,&e); camz = pSect->floorz;
 		if (pfbInsert(pIni, OBJ_WALL, s, nSect, camx, camy, pSect->floorz, 0) > 0)
 		{
@@ -2431,8 +2444,15 @@ static char getThumbnail_PFB(char* filepath, int nTile, int twh, int thg, int bg
 
 			//if (pfbAttachThumbnail(pIni, nTile)) // finally, attach it!
 				//pIni->Save();
-
-			hgltReset(kHgltPoint);
+			
+			if (pHglt)
+			{
+				for (p = pHglt->GetPtr(), highlightcnt = 0; *p >= 0; p++, highlightcnt++)
+					highlight[highlightcnt] = *p;
+			}
+			else
+				hgltReset(kHgltPoint);
+			
 			nRetn = 1;
 		}
 
@@ -2575,23 +2595,38 @@ static char getThumbnail_MAP(char* filepath, int nTile, int wh, int hg, int bg)
 				if ((pWalls = (walltype*)malloc(sizeof(walltype)*nWalls)) != NULL)
 					read(hFile, pWalls, sizeof(walltype)*nWalls); // that's it!
 				break;
+			case 5L:
 			case 6L:
 				read(hFile, &nSects, 2);
-				lseek(hFile, sizeof(SECTOR6)*nSects, SEEK_CUR);
+				lseek(hFile, sizeof(SECTOR6)*nSects, SEEK_CUR); // 5 and 6 are identical
 				read(hFile, &nWalls, 2);
 				if ((pWalls = (walltype*)malloc(sizeof(walltype)*nWalls)) != NULL)
 				{
 					for (i = 0; i < nWalls; i++)
 					{
-						WALL6 wall;
 						pWall = &pWalls[i];
-						read(hFile, &wall, sizeof(wall));
+						
+						if (ver == 5)
+						{
+							WALL5 wall;
+							read(hFile, &wall, sizeof(wall));
 
+							pWall->x 			= wall.x;
+							pWall->y 			= wall.y;
+							pWall->point2 		= wall.point2;
+							pWall->nextwall 	= wall.nextwall1;
+						}
+						else
+						{
+							WALL6 wall;
+							read(hFile, &wall, sizeof(wall));
 
-						pWall->x 			= wall.x;
-						pWall->y 			= wall.y;
-						pWall->point2 		= wall.point2;
-						pWall->nextwall 	= wall.nextwall;
+							pWall->x 			= wall.x;
+							pWall->y 			= wall.y;
+							pWall->point2 		= wall.point2;
+							pWall->nextwall 	= wall.nextwall;
+						}
+
 					}
 				}
 				break;
